pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "researchflow"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "requests",
    "google-genai",
    "python-dotenv",
    "httpx",
]

[project.scripts]
alarm = "researchflow.alarm.cli:main"
review = "researchflow.review.cli:main"